chore(srtripe): move stripe over to pages/account
